Update Player Class to generalize
Move game specific queries out of Player and into the unwritten "Game"
class. This way, we can do all the error checking in the game, where the
Game knows what input it is looking for. This way, the Player class is
not weirdly in the know about what it needs to do.